Remove non-generic DAP adapters from the list
The list of DAP adapters is highly problematic in many ways. The index
page writes:

> The Debug Adapter Protocol makes it possible to implement a generic
> debugger for a development tool that can communicate with different
> debuggers via Debug Adapters. And Debug Adapters can be re-used across
> multiple development tools which significantly reduces the effort to
> support a new debugger in different tools.

However, most of the "adapters" in this list are VSCode specific, and
worse yet several don't actually implement DAP in any reasonable respect.
This makes it very difficult for non-VSCode tools to actually use these
adapters which is antithetical to the ideas behind DAP.

The problematic items can be broken down into multiple categories:
1. Adapters which implement DAP, but which completely unspecified on how
   to run.

   Examples include "Microsoft/vscode-cpptools" which, to run, you must
   [copy a file and run another random file with
   node](https://github.com/mfussenegger/nvim-dap/wiki/C-C---Rust-(gdb-via--vscode-cpptools))

2. "adapters" which are simply VSCode plugin shims to an actual
   DAP adapter.

   Examples include "vscode-byebug" which wraps the "bybug-dap" adapter.

In this PR I have removed both, as they are entirely not useful for
non-VSCode tools.

However, many of these adapters still retain useful code. I am open to
alternative solutions on how to deal with these projects. One way to
deal with them would be to add a "VSCode plugin only" column to the
list of projects. Then users of non-VSCode tools would be able to still
find the list helpful.
Another way to deal with this would be to remove each of these projects
for now, but also leave them each an issue upstream on how they could be
re-added (e.g. by adding instructions on how to actually use the DAP
server)
Finally, the VSCode project could release a project to easily extract
the DAP adapters and language servers from its plugins.