update adapter table from community feedback
This commit adds a "Usage" column. If an extension in the table already
linked to a VS Code extension, I added that to the "Usage" column. For
those that did not, I tried to find appropriate usage docs. When
implementing a debug adapter, extension authors can use the "Usage"
column to link to any number of extensions/plugins or documentation.

I also expanded on the blurb before the table to clarify the purpose of
the page, and note that standalone debug adapters are possible. However
I did not manually go through and parse all repos to try to find other
usage instructions. Perhaps we tag all the maintainers if this commit
is merged so they can add any additional documentation they think is
appropriate.